org cem vector
